Move playerDead to its own file in game
